Organization_enum refactoring
